Add files via upload
This is my update on current Financial  Data that I am working with for a new Dashboard. The Data had already been cleaned and some elements are a bit difficult to work with. Unfortunately I could not find the original data and its source as I found it on an website through google search. However, it is meant to show visualizations for the Income Statement, Balance Sheet and Cashflow Statement for the Multi-Company corporation. I will have to update it once I am done but I shared my current progress